extract websockets into class
