+++ app.js
@@ -5,9 +5,11 @@ const server = require('http').Server(app);
 const io = require('socket.io')(server);
 const mongoose = require('mongoose');
 const config = require("./config/server");
+const WebSockets = require("./websockets");
 
 mongoose.Promise = Promise;
 
+const websockets = new WebSockets(io);
 // monter les routes
 const emission = require('./routes/emission.js');
 
@@ -22,40 +24,6 @@ app.use(bodyParser.urlencoded({ extended: false }))
 .use((err, req, res, next) => {
   if(err) return res.status(500).send({ error: err });
   next();
-});
-
-const CONNECTION = "connection",
-SET_TITLE = "setTitle",
-SET_INCRUST = "setIncrust",
-CLEAR_INCRUST = "clearIncrust",
-DEFAULT_TITLE = "GeekInc Remote Regie ready !";
-
-io.on(CONNECTION, socket => {
-
-  var config = {
-    text: DEFAULT_TITLE,
-    timeout: 2
-  };
-
-  socket.emit(SET_TITLE, config);
-
-  socket.on(SET_TITLE, data => {
-    console.log(data);
-    config = data; //TODO merge config & data
-    io.sockets.emit(SET_TITLE, data);
-  });
-
-  socket.on(SET_INCRUST, data => {
-    console.log(data);
-    //TODO check the file exists in the directory
-    io.sockets.emit(SET_INCRUST, data);
-  });
-
-  socket.on(CLEAR_INCRUST, data => {
-    console.log(data);
-    io.sockets.emit(CLEAR_INCRUST, data);
-  });
-
 });
